Roto, migrar a media2

+++ app/src/main/java/cobos/santiago/ui/componentes/musicComponents/HomeMusicScreen.kt
@@ -98,7 +98,7 @@ fun HomeContent(
                             MusicListItem(
                                 vm = vm,
                                 song = it,
-                                index = it.id
+                                index = it.id.toInt()
                             )
                         }
                     }
